Refactor: [FE] asset sub page에도 props를 하위 컴포넌트로 이관
